+++ functions/endpoints/affirmations/affirmations.js
@@ -1,8 +1,8 @@
-const { firestore, discordMessageEmbed } = require('../config');
-const { blaccSmithLogo } = require('../constants');
-const { randomIndex } = require('../utilities');
+const { firestore, discordMessageEmbed } = require('../../config');
+const { blaccSmithLogo } = require('../../constants');
+const { randomIndex } = require('../../utilities');
 
-module.exports = async () => {
+exports.postAffirmation = async () => {
 	const affirmations = await getAffirmations();
 	const randomAffirmation = affirmations[randomIndex(affirmations.length - 1)];
 	return embedMessage(randomAffirmation);
