+++ functions/endpoints/affirmations/index.js
