+++ functions/endpoints/dailyCC/constants.js
