+++ functions/endpoints/dailyCC/index.js
